feat: add about page
The code changes add a new About page to the project. This page displays a heading and a paragraph of text. This addition enhances the user experience by providing more information about the project or organization.

Note: The suggested commit message adheres to the established convention of using the "feat" prefix for new feature additions.